Add support for running certain tests without parallelization and ability to view tox output
Run tox without parallelization for all tox mps that have
parallel-tox == False. This could be due to waiting to avoid
race conditions when running tests in parallel. As is the case
for projects that use jenkins-job-builder.

Also added support for writing tox test output to file which is linked from report.