Running tox in a strictly confined snap will not work - help text added
Running tox in a strict snap is not possible as it tries to run external processes
using subprocess.Popen and also in lxc containers. This isn't possible with a
strictly confined snap so the parallelization work with snapcraft-preload will not work.

I've also added snap specific help text to the tox options noting that running tox
in a strictly confined snap will not work.